Transaction class now holds transaction amount in satoshi Removed hard coded exchange rate from various places throughout app (now uses ABC core) Added a bunch of TODOs in code so we don’t miss little things that need to get done Request controller now invokes QR Code display with most fields pre-populated (still need to get address in there but requires core change) QR code display hides tab bar while invoked

+++ AirBitz/Send/SendConfirmationViewController.h
@@ -15,7 +15,7 @@
 
 @property (assign) id<SendConfirmationViewControllerDelegate> delegate;
 @property (nonatomic, copy) NSString *sendToAddress;
-@property (nonatomic, assign) long long amountToSendSatoshi;
+@property (nonatomic, assign) int64_t amountToSendSatoshi;
 @property (nonatomic, assign) int selectedWalletIndex;
 @end
